requirements.txt
asgiref==3.5.2
Django==3.2.16
django-ckeditor==6.3.2
django-crispy-forms==1.14.0
django-js-asset==2.0.0
django-utils-six==2.0
gunicorn==20.1.0
Pillow==9.3.0
pytz==2022.6
sqlparse==0.4.3
typing-extensions==4.4.0
whitenoise==6.2.0
